fix line endings
Revert "fix line endings"

This reverts commit 0bc7b9e8639c79d19d3d3ba501023ce353ad9e02.

Fixed the framerate of 525 sequences, which were incorrectly set at 120fps, but should be 60.

fix line endings

Revert "Fixed the framerate of 525 sequences, which were incorrectly set at 120fps, but should be 60."

This reverts commit d046861da32e8b5ae704d245f472775d55f134db.

Fixed the framerate of 525 sequences, which were incorrectly set at 120fps, but should be 60.